requirements.txt
click==6.6
eventlet==0.19.0
Flask==0.11.1
Flask-Cors==3.0.2
Flask-PyMongo==0.4.1
Flask-SocketIO==2.7.1
greenlet==0.4.10
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
pymongo==3.3.0
python-engineio==1.0.3
python-socketio==1.5.1
six==1.10.0
Werkzeug==0.11.11

gcloud==0.18.1
grpcio==1.0.0
grpc-google-cloud-speech-v1beta1==1.0.1
